pyproject.toml
[tool.poetry]
name = "ammolite"
version = "0.1.1"
description = "Visualize structure data from Biotite with PyMOL"
readme = "README.rst"
license = "BSD-3-Clause"
authors = ["Patrick Kunzmann <patrick.kunzm@gmail.com>"]
homepage = "https://ammolite.biotite-python.org"
repository = "https://github.com/biotite-dev/ammolite"
documentation = "https://ammolite.biotite-python.org"
keywords = [
    "Bioinformatics",
    "Computational biology",
    "Data visualization"
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
packages = [
    { include = "ammolite", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.6"
biotite = "^0.21"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
sphinx = "^3.0"
numpydoc = ">= 0.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
